https://numworx.atlassian.net/browse/LMS-467
trash en undo activiteiten. initiele versie


git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/DWOProject/DWO-LMS/DWOJClient/trunk@25626 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/main/java/fi/dwo/dwojapplet/domain/Sco.java
@@ -88,7 +88,7 @@ public class Sco extends ScoBase implements LessonGroup, SCORM12APIInterface, Ap
     private Applet applet;
     private Boolean showScore;
     private byte[] imageData; // part of DomScoContextFull
-    private int sequencenr;
+    private long sequencenr;
 
     private boolean initialized;
 
@@ -779,7 +779,7 @@ public class Sco extends ScoBase implements LessonGroup, SCORM12APIInterface, Ap
      *
      * @return The sequencenumber of the sco.
      */
-    public int getSequencenr() {
+    public long getSequencenr() {
         return sequencenr;
     }
 
@@ -788,7 +788,7 @@ public class Sco extends ScoBase implements LessonGroup, SCORM12APIInterface, Ap
      *
      * @param sequencenr The sequencenumber to set.
      */
-    public void setSequencenr(int sequencenr) {
+    public void setSequencenr(long sequencenr) {
         if (!isCourseChanged()) {
             setCourseChanged(this.sequencenr != sequencenr);
         }
